streampos may be signed on some platforms

+++ src/support/file.cpp
@@ -29,12 +29,12 @@ T wasm::read_file(const std::string &filename, bool debug) {
   }
   infile.seekg(0, std::ios::end);
   std::streampos insize = infile.tellg();
-  if (insize >= std::numeric_limits<size_t>::max()) {
+  if (size_t(insize) >= std::numeric_limits<size_t>::max()) {
     // Building a 32-bit executable where size_t == 32 bits, we are not able to create strings larger than 2^32 bytes in length, so must abort here.
     std::cerr << "Failed opening '" << filename << "': Input file too large: " << insize << " bytes. Try rebuilding in 64-bit mode." << std::endl;
     exit(EXIT_FAILURE);
   }
-  T input((size_t)insize + 1, '\0');
+  T input(size_t(insize) + 1, '\0');
   infile.seekg(0);
   infile.read(&input[0], insize);
   return input;
